[Feature/store] 내 서랍(store) 2차 스프린트 기능구현 (#160)
* [feature/mypage]: add image

* [feature/mypage]: feat seekbar

* [feature/mypage]: chore code style

* [feature/mypage]: fix 15장 모두 찼을 때 글씨가 나오지 않는 현상 해결

* [feature/store]: refactor PR 1차 대응

* [feature/store]: migration for V2

* [feature/store]: DataSource 삭제, network dir 이동

* [feature/store]: refactor id Int -> Long

* [feature/share]: fix rollback 의진 코드

* [feature/share]: photo albumId Int to Long

* [feature/share]: photoLimit x프로퍼티 추가로 인한 대응

+++ app/src/main/java/com/teampophory/pophory/data/network/model/photo/PhotoRequest.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PhotoRequest(
     @SerialName("albumId")
-    val albumId: Int,
+    val albumId: Long,
     @SerialName("takenAt")
     val takenAt: String,
     @SerialName("studioId")
